Added parser.py
Added the file "parser.py". This file is intended to be used as a file parser for all input files that are given to the application. Parser.py would return a python dictionary which would contain a list of control points, knots, and degree value that would be translated later into a BSplineGeometry